Fix validation (#33)

+++ app/Http/Controllers/CategoryController.php
@@ -1,35 +1,43 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use App\Http\Requests\CategoryRequest;
+use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Validator;
+
 class CategoryController extends Controller
 {
     public function add(Request $request){
-        $request->validate([
-            'name' => ['required', 'max:40']
+        $validator=Validator::make($request->all(),[
+            'name' => 'required|string',
         ]);
+        if($validator->fails()){
+            return response()->json(['errors'=>$validator->errors()],422);
+        }
        Category::create([
            'name'=>$request->name,
        ]);
         return response()->json('Added Successfully');
     }
     public function edit(Request $request){
-        $request->validate([
-            'name' => ['required', 'max:40']
+        $validator=Validator::make($request->all(),[
+            'name' => 'required|string',
         ]);
-        $category=Category::findorfail($request->id);
-        $category->update([
-            'name'=>$request->name,
-        ]);
+        if($validator->fails()){
+            return response()->json(['errors'=>$validator->errors()],422);
+        }
+        $category=Category::find($request->id);
+        if(!$category){return response()->json(['error'=>'Category not found'],404);}
+        $category->update($request->all());
         return response()->json('Updated Successfully');
     }
-    public function delete(CategoryRequest $request){
-        $category=Category::findorfail($request->id);
-        $category->products()->delete();
-        $category->delete();
-        return response()->json('Deleted Successfully');
+    public function delete( $id){
+            $category=Category::find($id);
+            if(!$category){return response()->json(['error'=>'Category not found'],404);}
+            $category->products()->delete();
+            $category->delete();
+            return response()->json('Deleted Successfully');
+            
     }
     public function get(){
         $category=Category::all();
